feat: implement Sovereign Core Mechanical Truth expansion
- Purged all binary artifacts and enforced 100% source build via .gitignore.
- Fixed AHCI re-initialization loops and memory collisions with per-port CTBA buffers.
- Implemented mechanical FAT cluster walking in `vfs_read` for large file support.
- Hardened FAT32 persistence: fixed cluster alignment math and added `f_close` metadata syncing.
- Integrated XHCI BIOS Handover and enforced SATA-over-ROM boot priority.
- Implemented global Safe Mode in VFS triggered by mount/discovery failures.
- Updated architecture documentation to reflect the finalized 100% mechanical system.

+++ kernel/libs/fatfs/ff.h
@@ -9,7 +9,7 @@ typedef uint32_t DWORD;
 typedef uint16_t WORD;
 typedef uint8_t BYTE;
 typedef uint32_t FSIZE_t;
-typedef uint32_t LBA_t;
+typedef uint64_t LBA_t;
 
 typedef enum {
     FR_OK = 0,
